refactor the mess and divide into models and controllers
