+++ models/user.js
@@ -1,30 +1,6 @@
 'use strict'
 
 const db = require('../db');
-const bcrypt = require('bcrypt');
-
-/**
- *  Validate user data before register.
- * @param {{username: String, password: String, email: String }} data 
- */
-
-const validate = data => {
-    const usernameMatcher =  /^[a-z0-9_-]{3,16}$/;
-    const passwordMatcher = /^(?=.*\d)(?=.*[a-z])[0-9a-zA-Z]{8,}$/; 
-    if (usernameMatcher.test(data.username) &&
-        passwordMatcher.test(data.password)) {
-            return Object.assign({}, data);
-    }
-    else {
-        throw new Error("Registration Error");
-    }
-}
-
-const hashPassword = password => {
-    const salt = 10;
-    const hash = bcrypt.hash(password, salt);
-    return hash;
-}
 
 class User {
 
@@ -35,52 +11,7 @@ class User {
         this.email = email;
         this.spotify = spotify;
     }
-
-    /**
- 	* Authorization middleware.
- 	*/
-
-    static authorize(){
-        return async function(ctx, next) {
-            const data = ctx.request.body;
-            const user = await User.selectOneByUsername(data.username);
-
-            if (user === undefined) { throw new Error("Authorization Error"); }
-            const passwordValidation = await bcrypt.compare(data.password, user.password);
-            if (!passwordValidation) { throw new Error("Authorization Error"); }
     
-            ctx.session.userid = user.userid;
-    		ctx.session.authorized = true;
-
-            await next()
-     	  }
-	}
-	
-	/**
- 	* Registration middleware.
-	 */
-
-    static register(){ 
-        return async (ctx, next) => {
-            const data = validate(ctx.request.body);
-            data.password = await hashPassword(data.password); 
-            await db.query(db.sql.insert("User", data));
-            await next();
-        }      
-    }
-
-    /**
- 	* Logout middleware.
- 	*/
-
-    static logout(){
-        return async (ctx, next) => {
-            ctx.session.userid = undefined;
-        	ctx.session.authorized = false;
-            await next();
-        }
-    }
-
 	/**
  	* 
  	* @param {String} userid
